refactoring: extracting api call to a separate file & renaming files
